Dispara en direccion

+++ Projectile.h
@@ -3,22 +3,22 @@
 #define PROJECTILE_H
 
 #include "Map.h"
-#include <gtk/gtk.h>  // Añadir para usar GtkWidget
+#include <gtk/gtk.h>
 
 class Projectile {
 public:
-    int x, y; // Posición del proyectil
-    double direction_x, direction_y; // Dirección del proyectil
-    double speed; // Velocidad del proyectil
+    double x, y;  // Usamos double para mayor precisión en la dirección
+    double direction_x, direction_y; // Dirección normalizada del proyectil
+    double speed;  // Velocidad del proyectil
     int map_width; // Anchura del mapa
-    GtkWidget* widget; // Widget para dibujar el proyectil
+    GtkWidget* widget;  // Widget para el proyectil
     bool active = true;
 
-    Projectile(int start_x, int start_y, double dir_x, double dir_y, double spd, int map_width)
+    // Constructor
+    Projectile(double start_x, double start_y, double dir_x, double dir_y, double spd, int map_width)
     : x(start_x), y(start_y), direction_x(dir_x), direction_y(dir_y), speed(spd), map_width(map_width), widget(nullptr) {}
 
-
-    // Declaración del método update (la implementación irá en Projectile.cpp)
+    // Método para actualizar la posición del proyectil
     void update();
 };
